Fractals: extract shared df64 deep-zoom module (lib/df64.ts)
The emulated-double (df64) deep-zoom shader was copy-pasted between FractalsGPU and
Correspondence/FractalPane — byte-identical dfAdd/dfMul/dfNeg helpers, with MAX_ITER
4000 hardcoded in four places and splitDouble/suggestedIter duplicated. A precision
fix had to be mirrored into both; the drift risk was real.

Extract DF64_GLSL (the error-free-transform primitives + MAX_ITER) and the JS helpers
(splitDouble, MAX_ITERATIONS, suggestedIter) into lib/df64.ts, inlined into both shaders
the same way PALETTE_GLSL already is. Now a df64 fix lands in one place.

Verified: build + lint (0 errors), df64-gpu-probe passes, and both #/fractals and
#/correspondence render correctly (shaders compile with the inlined DF64_GLSL).
